fix: update terminator module name and regenerate type stubs
